requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-admin-honeypot==1.1.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==3.9.0
django-filter==21.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-notifications-rest==0.0.1
django-phonenumber-field==5.2.0
django-storages==1.11.1
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.1
jsonfield==3.1.0
kombu==5.1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
phonenumbers==8.12.33
Pillow==8.3.2
prompt-toolkit==3.0.20
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==21.0.0
python-crontab==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
swapper==1.1.2.post1
twilio==7.0.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zope.interface==5.4.0
